backend/requirements.txt
# Django Core
Django>=4.2,<5.0
djangorestframework>=3.14.0

# JWT Authentication
djangorestframework-simplejwt>=5.3.0

# CORS Headers
django-cors-headers>=4.3.0

# Production
gunicorn>=21.0.0
python-dotenv>=1.0.0
